+++ crates/ryo3-which/src/lib.rs
@@ -17,12 +17,12 @@ use pyo3::prelude::*;
 
 #[pyfunction]
 #[pyo3(signature= (cmd, path=None))]
-pub fn which(cmd: &str, path: Option<&str>) -> PyResult<Option<PathBuf>> {
+pub fn which(py: Python<'_>, cmd: &str, path: Option<&str>) -> PyResult<Option<PathBuf>> {
     if let Some(p) = path {
         // get current directory w/o unwrapping
         match env::current_dir() {
             Ok(c) => {
-                let which_res = ::which::which_in(cmd, Some(p), c);
+                let which_res = py.detach(|| which_rs::which_in(cmd, Some(p), c));
                 match which_res {
                     Ok(p) => Ok(Some(p)),
                     Err(_e) => Ok(None),
@@ -33,7 +33,7 @@ pub fn which(cmd: &str, path: Option<&str>) -> PyResult<Option<PathBuf>> {
             )),
         }
     } else {
-        let r = which_rs::which(cmd);
+        let r = py.detach(|| which_rs::which(cmd));
         match r {
             Ok(p) => Ok(Some(p)),
             Err(_e) => Ok(None),
@@ -43,19 +43,21 @@ pub fn which(cmd: &str, path: Option<&str>) -> PyResult<Option<PathBuf>> {
 
 #[pyfunction]
 #[pyo3(signature= (cmd, path=None))]
-pub fn which_all(cmd: &str, path: Option<&str>) -> PyResult<Vec<PathBuf>> {
+pub fn which_all(py: Python<'_>, cmd: &str, path: Option<&str>) -> PyResult<Vec<PathBuf>> {
     let search_path: Option<OsString> = match path {
         Some(p) => Some(OsString::from(p)),
         None => env::var_os("PATH"),
     };
     let curdir = env::current_dir()?;
-    let which_iter = which_rs::which_in_all(cmd, search_path, curdir).map_err(|_| {
+    py.detach(|| {
+        which_rs::which_in_all(cmd, search_path, curdir)
+            .map(|p| p.into_iter().collect::<Vec<PathBuf>>())
+    })
+    .map_err(|_| {
         PyErr::new::<pyo3::exceptions::PyOSError, _>(
             "which_all: current directory is not a valid path",
         )
-    })?;
-    let which_vec = which_iter.into_iter().collect::<Vec<_>>();
-    Ok(which_vec)
+    })
 }
 
 pub fn pymod_add(m: &Bound<'_, PyModule>) -> PyResult<()> {